+++ src/rsg-components/Logo/Logo.spec.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import LogoRenderer from './LogoRenderer';
+
+it('renderer should render header', () => {
+	const actual = render(<LogoRenderer>React Styleguidist</LogoRenderer>);
+
+	expect(actual).toMatchSnapshot();
+});